flstudio: Separate into flstudio and flstudio_core
This is so the minimal amount of code can be licensed GPL v3 as
possible.
